requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
instaloader>=4.13.0
playwright>=1.48.0
jinja2>=3.1.4
pypdf>=5.0.0
pillow>=10.4.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
pydantic>=2.9.0
